refactor: rename rct_analysis.py to rct_utils.py, remove detect_design_dimension
Addresses PR review feedback:
- Rename stgym/rct_analysis.py → stgym/rct_utils.py
- Remove detect_design_dimension() (no longer needed — design dimension
  is read directly from run.data.tags.design_dimension)
- Update all imports accordingly
