+++ tests/test_marimo_notebook.py
@@ -53,7 +53,7 @@ _POOLING_DIM = "model.pooling.type"
 _POOLING_PARAM = "model/mp_layers/0/pooling/type"
 
 
-@mock.patch("stgym.rct_analysis.fetch_runs")
+@mock.patch("stgym.rct_utils.fetch_runs")
 class TestAnalyzeExperimentWithRealisticData:
     """Test analyze_experiment with realistic mock data including k-fold runs."""
 
@@ -126,7 +126,7 @@ class TestAnalyzeExperimentWithRealisticData:
 
     def test_with_regular_runs_only(self, mock_fetch_runs: MagicMock):
         """Verify analysis works with regular (non-k-fold) runs."""
-        from stgym.rct_analysis import analyze_experiment
+        from stgym.rct_utils import analyze_experiment
 
         mock_fetch_runs.return_value = self.regular_runs
 
@@ -148,7 +148,7 @@ class TestAnalyzeExperimentWithRealisticData:
 
     def test_with_kfold_runs_only(self, mock_fetch_runs: MagicMock):
         """Verify analysis correctly aggregates k-fold runs by mean metric."""
-        from stgym.rct_analysis import analyze_experiment
+        from stgym.rct_utils import analyze_experiment
 
         mock_fetch_runs.return_value = self.kfold_runs
 
@@ -168,7 +168,7 @@ class TestAnalyzeExperimentWithRealisticData:
 
     def test_with_mixed_runs(self, mock_fetch_runs: MagicMock):
         """Verify analysis handles mixed regular and k-fold runs."""
-        from stgym.rct_analysis import analyze_experiment
+        from stgym.rct_utils import analyze_experiment
 
         mock_fetch_runs.return_value = self.mixed_runs
 
@@ -187,7 +187,7 @@ class TestAnalyzeExperimentWithRealisticData:
 
     def test_kfold_aggregation_computes_mean(self, mock_fetch_runs: MagicMock):
         """Verify k-fold aggregation computes correct mean metrics."""
-        from stgym.rct_analysis import (
+        from stgym.rct_utils import (
             aggregate_kfold_metrics,
             runs_to_dataframe,
         )
@@ -210,7 +210,7 @@ class TestAnalyzeExperimentWithRealisticData:
 
     def test_failed_fold_marks_group_failed(self, mock_fetch_runs: MagicMock):
         """Verify a failed fold marks the entire aggregated group as failed."""
-        from stgym.rct_analysis import (
+        from stgym.rct_utils import (
             aggregate_kfold_metrics,
             filter_complete_groups,
             runs_to_dataframe,
@@ -231,7 +231,7 @@ class TestAnalyzeExperimentWithRealisticData:
 
     def test_rank_summary_with_kfold_data(self, mock_fetch_runs: MagicMock):
         """Verify rank summary statistics work with k-fold aggregated data."""
-        from stgym.rct_analysis import (
+        from stgym.rct_utils import (
             analyze_experiment,
             summarize_ranks_by_design_choice,
         )
